requirements.txt
Django==4.2.6
pillow==11.1.0
whitenoise==6.8.2
python-dotenv==1.0.1
